+++ src/main/kotlin/alvarx4z/medialibrary/domain/book/PageCount.kt
@@ -1,5 +1,5 @@
 package alvarx4z.medialibrary.domain.book
 
-import alvarx4z.medialibrary.domain.shared.invariant.PositiveInt
+import alvarx4z.medialibrary.domain.shared.invariants.PositiveInt
 
 class PageCount(value: Int) : PositiveInt(value)
