+++ src/main/kotlin/alvarx4z/medialibrary/domain/book/PublicationDate.kt
@@ -1,6 +1,6 @@
 package alvarx4z.medialibrary.domain.book
 
-import alvarx4z.medialibrary.domain.shared.invariant.PastDate
+import alvarx4z.medialibrary.domain.shared.invariants.PastDate
 import java.time.LocalDate
 
 class PublicationDate(date: LocalDate) : PastDate(date)
