+++ src/main/kotlin/alvarx4z/medialibrary/domain/book/series/Order.kt
@@ -1,5 +1,5 @@
 package alvarx4z.medialibrary.domain.book.series
 
-import alvarx4z.medialibrary.domain.shared.invariant.PositiveInt
+import alvarx4z.medialibrary.domain.shared.invariants.PositiveInt
 
 class Order(value: Int) : PositiveInt(value)
