+++ src/main/kotlin/alvarx4z/medialibrary/domain/shared/Summary.kt
@@ -1,5 +1,5 @@
 package alvarx4z.medialibrary.domain.shared
 
-import alvarx4z.medialibrary.domain.shared.invariant.NotEmptyString
+import alvarx4z.medialibrary.domain.shared.invariants.NotEmptyString
 
 data class Summary(val resource: NotEmptyString)
