+++ src/main/kotlin/alvarx4z/medialibrary/domain/shared/invariants/NotEmptyString.kt
@@ -1,4 +1,4 @@
-package alvarx4z.medialibrary.domain.shared.invariant
+package alvarx4z.medialibrary.domain.shared.invariants
 
 data class NotEmptyString(val value: String) {
   init {
