+++ src/main/kotlin/alvarx4z/medialibrary/domain/shared/professional/Date.kt
@@ -1,6 +1,6 @@
 package alvarx4z.medialibrary.domain.shared.professional
 
-import alvarx4z.medialibrary.domain.shared.invariant.PastDate
+import alvarx4z.medialibrary.domain.shared.invariants.PastDate
 import java.time.LocalDate
 
 class Date(date: LocalDate) : PastDate(date)
