+++ src/main/kotlin/alvarx4z/medialibrary/domain/shared/professional/Name.kt
@@ -1,6 +1,6 @@
 package alvarx4z.medialibrary.domain.shared.professional
 
-import alvarx4z.medialibrary.domain.shared.invariant.NotEmptyString
+import alvarx4z.medialibrary.domain.shared.invariants.NotEmptyString
 
 data class Name(
   val name: NotEmptyString,
