+++ src/test/kotlin/alvarx4z/medialibrary/domain/book/ISBNTest.kt
@@ -1,7 +1,7 @@
 package alvarx4z.medialibrary.domain.book
 
 import alvarx4z.medialibrary.BookHelper
-import alvarx4z.medialibrary.domain.shared.invariant.NotEmptyString
+import alvarx4z.medialibrary.domain.shared.invariants.NotEmptyString
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
