+++ src/test/kotlin/alvarx4z/medialibrary/domain/shared/SummaryTest.kt
@@ -1,7 +1,7 @@
 package alvarx4z.medialibrary.domain.shared
 
 import alvarx4z.medialibrary.SharedHelper
-import alvarx4z.medialibrary.domain.shared.invariant.NotEmptyString
+import alvarx4z.medialibrary.domain.shared.invariants.NotEmptyString
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
